PIR & LED Print with Alarm, no twitter
Almost there.  This code reads whether a hand is in the jar or not
(using a Passive Infared Sensor) and sounds an alarm while writing "The
Cookies are in Danger" to the LCD screen.

In order to be able to tweet, I will need a wifi shield, which I don't
currently have.